feat(desktop, navigation): quick screen change for scroll
